requirements.txt
asgiref==3.8.1
bleach==6.1.0
Django==4.2.11
django-allauth==0.62.1
djangorestframework==3.15.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
webencodings==0.5.1
